+++ viadot/task_utils_v2.py
@@ -9,14 +9,13 @@ METADATA_COLUMNS = {"_viadot_downloaded_at_utc": "DATETIME"}
 
 @task
 def add_ingestion_metadata_task(
-    path: str,
+    df: pd.DataFrame,
 ):
     """Add ingestion metadata columns, eg. data download date
 
     Args:
-        path (str): The path to the CSV file containing the data.
-        sep (str, optional): The separator to use when loading the file into the DataFrame. Defaults to "\t".
+        df (pd.DataFrame): input DataFrame.
     """
-    df = pd.read_parquet(path)
-    df["_viadot_downloaded_at_utc"] = datetime.now(timezone.utc).replace(microsecond=0)
-    df.to_parquet(path, index=False)
+    df2 = df.copy(deep=True)
+    df2["_viadot_downloaded_at_utc"] = datetime.now(timezone.utc).replace(microsecond=0)
+    return df2